requirements-api.txt
# FastAPI & UVicorn
fastapi==0.111.0        # needs Pydantic ≥ 2.4
uvicorn[standard]==0.29.0

# Auth & HTTP
python-jose==3.3.0
httpx==0.27.0
python-multipart==0.0.9

# Azure SDKs (API only needs to upload/download PDFs if you keep blob calls here)
azure-identity==1.16.0
azure-storage-blob==12.20.0

# Template rendering
jinja2>=3.1,<4

# Pydantic pinning
pydantic>=2.7,<3
